use Balances components
